hw12: move config to separate package
